MikePY/python-service/requirements.txt


fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2
python-dotenv==1.0.0
pymongo==4.5.0
httpx==0.25.0
datasets==2.14.5
google-cloud-bigquery==3.12.0
tensorflow==2.13.0
pydantic==2.4.2
python-multipart==0.0.6